+++ DynamicArrays/DynamicArrays/FirstTasks.h
@@ -1,9 +1,9 @@
 #ifndef FIRSTTASKS_H
 #define FIRSTTASKS_H
 #include <iostream>
-int32_t MinElementIndex(int32_t*,int32_t);
-int32_t FindNegative(int32_t*, int32_t, bool);
-int32_t SummBetweenNegative(int32_t*, int32_t);
-void InputNumber(int32_t&);
-void ChangeArray(int32_t*, int32_t, int32_t);
+int32_t MinElementIndex(double*,int32_t);
+int32_t FindNegative(double*, int32_t, bool);
+double SummBetweenNegative(double*, int32_t);
+void InputNumber(double&);
+void ChangeArray(double*, int32_t, double);
 #endif
